orchestra/daemon: Convert to a subpackage
