+++ teuthology/orchestra/daemon/__init__.py
@@ -1,0 +1,1 @@
+from teuthology.orchestra.daemon.group import DaemonGroup  # noqa
